Add helper methods; fix contentLength type

+++ ktor-http/src/io/ktor/http/HttpMessageProperties.kt
@@ -18,14 +18,15 @@ fun HttpMessageBuilder.contentType(): ContentType? = headers[HttpHeaders.Content
 fun HttpMessageBuilder.charset(): Charset? = contentType()?.charset()
 fun HttpMessageBuilder.etag(): String? = headers[HttpHeaders.ETag]
 fun HttpMessageBuilder.vary(): List<String>? = headers[HttpHeaders.Vary]?.split(",")?.map { it.trim() }
-fun HttpMessageBuilder.contentLength(): Int? = headers[HttpHeaders.ContentLength]?.toInt()
+fun HttpMessageBuilder.contentLength(): Long? = headers[HttpHeaders.ContentLength]?.toLong()
 
 fun HttpMessage.contentType(): ContentType? = headers[HttpHeaders.ContentType]?.let { ContentType.parse(it) }
 fun HttpMessage.charset(): Charset? = contentType()?.charset()
 fun HttpMessage.etag(): String? = headers[HttpHeaders.ETag]
 fun HttpMessage.vary(): List<String>? = headers[HttpHeaders.Vary]?.split(",")?.map { it.trim() }
-fun HttpMessage.contentLength(): Int? = headers[HttpHeaders.ContentLength]?.toInt()
-fun HttpMessage.setCookie(): List<Cookie> = headers.getAll(HttpHeaders.SetCookie)?.map { parseServerSetCookieHeader(it) } ?: emptyList()
+fun HttpMessage.contentLength(): Long? = headers[HttpHeaders.ContentLength]?.toLong()
+fun HttpMessage.setCookie(): List<Cookie> =
+    headers.getAll(HttpHeaders.SetCookie)?.map { parseServerSetCookieHeader(it) } ?: emptyList()
 
 fun HttpMessageBuilder.cookies(): List<Cookie> =
-        headers.getAll(HttpHeaders.SetCookie)?.map { parseServerSetCookieHeader(it) } ?: emptyList()
+    headers.getAll(HttpHeaders.SetCookie)?.map { parseServerSetCookieHeader(it) } ?: emptyList()
